+++ Cargo.toml
@@ -17,10 +17,13 @@ exclude = ["tests/pocs/blob"]
 members = [".", "./src/proc-macros", "./wasm"]
 
 [features]
-default = ["mmap", "derive_pod"]
+default = []
 unsafe_alignment = []
-mmap = ["std", "libc", "winapi"]
+mmap = ["std", "alloc", "libc", "winapi"]
 unstable = []
+pattern = []
+scanner = []
+alloc = ["hashbrown"]
 derive_pod = ["dataview/derive_pod"]
 std = ["no-std-compat/std"]
 resources_nostd = ["hashbrown"]
@@ -39,12 +42,13 @@ all-features = true
 default-target = "x86_64-pc-windows-msvc"
 
 [dependencies]
+heapless = { version = "0.9.3" }
 pelite-macros = { path = "src/proc-macros", version = "0.1.1" }
 dataview = { version = "1.0", default-features = false }
 serde = { version = "1.0", optional = true, features = ["derive"] }
 data-encoding = { version = "2.3", optional = true }
-no-std-compat = { version = "0.4.0", features = ["alloc"] }
-hashbrown = { version = "0.8.0", optional = true }
+no-std-compat = { version = "0.4.0", optional = true, features = ["alloc"] }
+hashbrown = { version = "0.17.1", optional = true }
 
 [target.'cfg(unix)'.dependencies]
 libc = { version = "0.2", optional = true }
