+++ examples/apisetschema/win10.rs
@@ -37,6 +37,7 @@ impl<'a> Schema<'a> {
 	pub fn header(self) -> &'a API_SET_NAMESPACE_V6 {
 		unsafe { &*(self.image.as_ptr() as *const API_SET_NAMESPACE_V6) }
 	}
+	
 	pub fn entries(&self) -> Result<Entries<'a>> {
 		let header = self.header();
 		let entries = self.slice_len(header.EntryOffset, header.Count as usize)?;
@@ -62,24 +63,31 @@ pub struct Entries<'a> {
 	entries: &'a [API_SET_NAMESPACE_ENTRY],
 	hashes: &'a [API_SET_HASH_ENTRY],
 }
+
 impl<'a> Entries<'a> {
+
 	pub fn entries(&self) -> &'a [API_SET_NAMESPACE_ENTRY] {
 		self.entries
 	}
+
 	pub fn hashes(&self) -> &'a [API_SET_HASH_ENTRY] {
 		self.hashes
 	}
+
 	pub fn iter(&self) -> impl Iterator<Item = Entry<'a>> + Clone + 'a {
 		let schema = self.schema;
 		self.entries.iter().map(move |image| Entry { schema, image })
 	}
+
 	pub fn len(&self) -> usize {
 		self.schema.header().Count as usize
 	}
+
 	pub fn at(&self, i: usize) -> Option<Entry<'a>> {
 		self.entries.get(i).map(|image| Entry { schema: self.schema, image })
 	}
 }
+
 #[rustfmt::skip]
 impl<'a> fmt::Debug for Entries<'a> {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -96,18 +104,23 @@ pub struct Entry<'a> {
 	schema: Schema<'a>,
 	image: &'a API_SET_NAMESPACE_ENTRY,
 }
+
 impl<'a> Entry<'a> {
+
 	pub fn is_sealed(&self) -> bool {
 		self.image.Flags & 1 != 0
 	}
+
 	pub fn name(&self) -> Result<&'a [u16]> {
 		self.schema.slice_len(self.image.NameOffset, (self.image.NameLength / 2) as usize)
 	}
+
 	pub fn values(&self) -> Result<Values<'a>> {
 		let image = self.schema.slice_len(self.image.ValueOffset, self.image.ValueCount as usize)?;
 		Ok(Values { schema: self.schema, image })
 	}
 }
+
 #[rustfmt::skip]
 impl<'a> fmt::Debug for Entry<'a> {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -126,21 +139,28 @@ pub struct Values<'a> {
 	schema: Schema<'a>,
 	image: &'a [API_SET_VALUE_ENTRY],
 }
+
 impl<'a> Values<'a> {
+
 	pub fn image(&self) -> &'a [API_SET_VALUE_ENTRY] {
 		self.image
 	}
+
 	pub fn iter(&self) -> impl Iterator<Item = Value<'a>> + Clone + 'a {
 		let schema = self.schema;
 		self.image.iter().map(move |image| Value { schema, image })
 	}
+
 	pub fn len(&self) -> usize {
 		self.image.len()
 	}
+
 	pub fn at(&self, i: usize) -> Option<Value<'a>> {
 		self.image.get(i).map(|image| Value { schema: self.schema, image })
 	}
+
 }
+
 #[rustfmt::skip]
 impl<'a> fmt::Debug for Values<'a> {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -157,17 +177,22 @@ pub struct Value<'a> {
 	schema: Schema<'a>,
 	image: &'a API_SET_VALUE_ENTRY,
 }
+
 impl<'a> Value<'a> {
+
 	pub fn is_sealed(&self) -> bool {
 		self.image.Flags & 1 != 0
 	}
+
 	pub fn api_set_name(&self) -> Result<&'a [u16]> {
 		self.schema.slice_len(self.image.NameOffset, (self.image.NameLength / 2) as usize)
 	}
+
 	pub fn host_name(&self) -> Result<&'a [u16]> {
 		self.schema.slice_len(self.image.ValueOffset, (self.image.ValueLength / 2) as usize)
 	}
 }
+
 #[rustfmt::skip]
 impl<'a> fmt::Debug for Value<'a> {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
