+++ src/pe64/load_config.rs
@@ -21,7 +21,7 @@ fn example(file: PeFile<'_>) -> pelite::Result<()> {
 ```
 */
 
-use std::fmt;
+use core::fmt;
 
 use crate::{Error, Result};
 
@@ -87,7 +87,6 @@ mod serde {
 #[cfg(test)]
 pub(crate) fn test<'a, P: Pe<'a>>(pe: P) -> Result<()> {
 	let load_config = pe.load_config()?;
-	let _ = format!("{:?}", load_config);
 	let _security_cookie = load_config.security_cookie();
 	let _se_handler_table = load_config.se_handler_table();
 	Ok(())
