+++ src/pe64/pe.rs
@@ -2,7 +2,7 @@
 Abstract over mapped images and file binaries.
 */
 
-use std::{cmp, mem, ptr, slice};
+use core::{cmp, mem, ptr, slice};
 
 use crate::{util::AlignTo, util::CStr, util::FromBytes, Pod};
 use crate::{Error, Result};
@@ -560,7 +560,8 @@ pub unsafe trait Pe<'a>: PeObject<'a> + Copy {
 	/// See the [resources](resources/index.html) module for more information.
 	///
 	/// Returns [`Err(Null)`](../enum.Error.html#variant.Null) if the image has no resources. Any other error indicates some form of corruption.
-	#[cfg(any(feature = "std", feature = "resources_nostd"))]
+	// #[cfg(any(feature = "std", feature = "resources_nostd"))]
+	#[cfg(all(feature = "alloc", feature = "std"))]
 	fn resources(self) -> Result<crate::resources::Resources<'a>>
 	where
 		Self: Copy,
@@ -574,6 +575,7 @@ pub unsafe trait Pe<'a>: PeObject<'a> + Copy {
 	/// Gets Scanner access.
 	///
 	/// See the [scanner](scanner/index.html) module for more information.
+	#[cfg(feature = "scanner")]
 	fn scanner(self) -> super::scanner::Scanner<Self> {
 		super::scanner::Scanner::new(self)
 	}
@@ -605,7 +607,7 @@ unsafe impl<'s, 'a> Pe<'a> for &'s dyn PeObject<'a> {}
 //----------------------------------------------------------------
 
 #[cfg(feature = "serde")]
-pub(crate) fn serialize_pe<'a, P: Pe<'a>, S: serde::Serializer>(pe: P, serializer: S) -> std::result::Result<S::Ok, S::Error> {
+pub(crate) fn serialize_pe<'a, P: Pe<'a>, S: serde::Serializer>(pe: P, serializer: S) -> core::result::Result<S::Ok, S::Error> {
 	use crate::util::serde_helper::*;
 
 	let mut state = serializer.serialize_struct(pe.serde_name(), 10)?;
