+++ tests/patterns.rs
@@ -1,3 +1,4 @@
+
 use pelite::pattern as pat;
 
 #[allow(dead_code)]
